Downgrade phantomjs to 1.9.8.
The old combination of phantomjs 2.1.1 + casperjs 1.1.4-1 was super
buggy.  The selftest crashed after reporting hundreds of failures.

This combination of phantomjs + casperjs is as good as it gets:

In the docker image:
casperjs selftest:

FAIL 1250 tests executed in 36.168s, 1249 passed, 1 failed, 0 dubious, 5 skipped.

Details for the 1 failed test:

In /casperjs-1.1.4-1/tests/suites/casper/wait.js:2512
  waitForExec() tests